task(refactor-and-fixes): Basic Cleanup
Move all projects to use a project.json. This is a cleaner approach in
my opinion and a workspace.json can get cluttered.

Move the rollup dependencies out of the local package.json of the sdk
and to the root pacakge.json. This is to help enforce the Single Version
Policy. Monorepos can run into problems relating to dependencies (as any
repo can) and this is how companies like Google manage versions to avoid
the Diamond Dependency Problem.

Created a small 'event' package to work on cleanup for how the jest
configs will operate and co-exist. Updated TS configs as needed and
updated references for 'event'