feature(E2E): Run E2E through Nx
In order to run e2e's through nx we need to define an 'app'. We don't
have a clear 'app' per se but I split the 'e2e' app into its own
'package'. This allows us to create a playwright e2e for the app itself.

Hypothetically, we could create a lot of individual apps and have a
potentially more affected build through nx. This was a big chunk of work
and I deem that to be a long term goal if possible. This still utilizes
webpack to run the tests.

This PR is probably not done, I need to do more testing around the
output folders to ensure its a stable and clean build/teardown as I
think there may be some outlier with ts outputs happening