task(move-env-for-samples): Leverage ENV in workspace Root
Moving the `env` file to the workspace root for sample apps. Nx by
default will use this as the *least* specific and therefore lowest
priority env file. Overrides to this can be done in the folder closer to
where the file is used per Nx documentation.

This allows React and Angular sample apps to share one env file and we
have no conflicts as a result