task(CI): Fix CI error
There was a breaking of the CI build during a clean install. There were
packages I should have cleaned up previously and I didn't.

Additionally, I added engine fields to the package.json so we are always
on the correct versions and updated CI builds to be the same across
different stages