feature(swc): Migrate things that can be swc to swc
Use the swc compiler. It is much faster than babel. Also move the
playwright config to be local to the e2e app. This ensures we can have a
more specialized config on an app basis. Also switched to 2 workers
because I think 4 was too fast for the node api server.