docs(ReadMe): Readme refactor
This is mostly a refactor of the readme.

This also refactors the centralized and embedded login samples into
their own bundled applications.

This simplifies the configuration for rollup, and allows the sdk to be
linked via webpack directly by the applications.

The readme also will map to the SDK/sample in the given directory

This address comments on the PR, we aligned the root env to essentially
be a list of all the existing env variables but have a more specific env
file in each sample app thats catered to that application.

Aligned scripts to follow the root npm script for running apps in read
me files.

The document was loading the element faster than the network request
could process so we were ending up in a flaky situation where the test
could fail.

As a result we have added a fix where we have a query param appended to
the url of the redirect in central login so that we have not loaded the
dom element we wait for until we actually have the authorized redirect