tests(add-mobile-devices): Track coverage by default and add Mobile devices
This adds code coverage report generation by default, uploads to code
cov and also implements a  lot more devices for us to test.

task(remove-nx-cloud-step): This step should be fixed

This was a temporary fix from a few weeks ago that can be removed now.

Added the caching step to the agents so that if we don't need to install
deps we shouldn't in each agent as well.