ci: automate-releases
we now can automate next and official releases. github events are
localized to the vm that they are run on. This means that the previous
workflows would never fully automate the process. now we run a merge job
that will eventually kick off a successful beta release to github and
npm.

official releases are only run on tags that start with a `v`, while the
beta releases ignore all tag pushes.