fix(javascript-sdk): fix-profile-object-checks
The device metadata collection code does feature detection to see if it
is safe to use a certain browser API without a crash. However, many of
the checks fail if the object it is checking is undefined. This commit
fixes that, by ensuring the typeof the object !== "undefined" (which is
always safe).

It also replaces the two usages of window.crypto with globalThis.crypto,
as that will work in both the browser and Node.js (even though Node
isn't officially supported, I don't think it should fall over when it
provides the same API in a different place).