refactor: ci-and-readme-changes
I felt that the merge CI pipeline was redundant. Merges by in large are
tested in the PR branch as that is the commit that is "run" by the
pipeline. I wanted to remove the testing that is redundant. I left
builds on and added the base tag to the pipeline for releases.

I fixed a dead link in the readme and then I removed a stale package
from the js-sdk repo package.json. I fixed the link in CONTRIBUTING.md
to also reflect the conventional commits link.