ci: angualar-executor-doesnt-have-head-and-sha
The angular executor does not have a head / sha argument when building.
This means that building based on the head and sha difference doesn't
work properly for angular. Its not super important for the build step on
merge, which is more of a sanity check than anything.

So now CI merges will just build the resulting merge (which happens in
the PR step anyways) and then will attempt a publish step