ci: add-luis-as-publisher
long story short, we need a user to make commits, sign them and be
allowed to push to protected branches in CI so that we can version and
update the package.json. Luis is easiest to do this as he's an admin,
and we don't need red tape of adding a bot user with liscenses.