chore: remove-version
release process added beta version, and then failed because of two
reasons. first the npm key was removed by me earlier in the day by
accident, and then the second time the version was pushed and the job
"re run" failed becaue the tip of the repo was ahead of the run.