test: update-token-vault-suites-and-release
token vault suites are fixed now and the tests should pass and run fine.

update a local publish script so that it spins up verdaccio and then uses

nx release to release all packages/* to verdaccio

for sanity sake - it tags as beta and publishes under that tag

it doesn't do any versioning or commiting. Just publish packages to verdaccio