chore: remove-project-json
This removes all instances of a project.json file.
This is just for demonstration purposes. This type
of setup uses `nx` more as a background task
and uses `pnpm` to execute scripts. Its
not very different than the existing setup

the core difference is you must reference
a project based on the name in its
`package.json` file rather than the short-name
that the project.json defined.